i think i finish christmastree
git-svn-id: https://version-control.adelaide.edu.au/svn/a1691807/2017/s2@845 ebee28f9-b968-474b-8da9-e5d0b78a977b

+++ pssd/week02practice/AlternateColors.hpp
@@ -35,7 +35,7 @@ class AlternateColors{
         }
 
         int currColor(long r,long g,long b,long k,int help){
-            int min = 0;
+            long min = 0;
             if(r<=g&&r<=b){
                 min = r;
                 help = 1;
